refactor: rename packages for group 'Concurrency'
