+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/common/Example1.java
@@ -1,14 +1,12 @@
-package com.rakovets.course.javabasics.example.concurrency;
+package com.rakovets.course.javabasics.example.concurrency.common;
 
-import com.rakovets.course.javabasics.example.concurrency.model.PrintRunnable;
 import com.rakovets.course.javabasics.util.AnsiColorCode;
 import com.rakovets.course.javabasics.util.StandardOutputUtil;
 
-public class Example3 {
+public class Example1 {
     public static void main(String[] args) {
         StandardOutputUtil.printlnWithTime("Main thread: started", AnsiColorCode.FG_RED_BOLD);
-        Runnable printRunnable = new PrintRunnable();
-        Thread thread = new Thread(printRunnable);
+        Thread thread = new Thread();
         thread.start();
         StandardOutputUtil.printlnWithTime("Main thread: finished", AnsiColorCode.FG_RED_BOLD);
     }
