+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/common/Example10_1.java
@@ -1,29 +1,28 @@
-package com.rakovets.course.javabasics.example.concurrency;
+package com.rakovets.course.javabasics.example.concurrency.common;
 
 import com.rakovets.course.javabasics.util.AnsiColorCode;
 import com.rakovets.course.javabasics.util.StandardOutputUtil;
 
-public class Example9_1 {
+public class Example10_1 {
     public static void main(String[] args) {
         StandardOutputUtil.printlnWithTimeAndThread("started", AnsiColorCode.FG_RED_BOLD);
 
         Runnable runnable = () -> {
-            while (true) {
+            for (int i = 0; i < 5; i++) {
                 StandardOutputUtil.printlnWithTimeAndThread("...", AnsiColorCode.FG_BLUE_BOLD);
-                sleep(1);
+                sleep(1000);
             }
         };
         Thread thread = new Thread(runnable);
+        thread.setDaemon(true);
         thread.start();
-
-        sleep(5);
 
         StandardOutputUtil.printlnWithTimeAndThread("finished", AnsiColorCode.FG_RED_BOLD);
     }
 
-    private static void sleep(long seconds) {
+    private static void sleep(long millis) {
         try {
-            Thread.sleep(seconds * 1000);
+            Thread.sleep(millis);
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
