+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/common/Example3.java
@@ -1,12 +1,14 @@
-package com.rakovets.course.javabasics.example.concurrency;
+package com.rakovets.course.javabasics.example.concurrency.common;
 
+import com.rakovets.course.javabasics.example.concurrency.common.model.PrintRunnable;
 import com.rakovets.course.javabasics.util.AnsiColorCode;
 import com.rakovets.course.javabasics.util.StandardOutputUtil;
 
-public class Example1 {
+public class Example3 {
     public static void main(String[] args) {
         StandardOutputUtil.printlnWithTime("Main thread: started", AnsiColorCode.FG_RED_BOLD);
-        Thread thread = new Thread();
+        Runnable printRunnable = new PrintRunnable();
+        Thread thread = new Thread(printRunnable);
         thread.start();
         StandardOutputUtil.printlnWithTime("Main thread: finished", AnsiColorCode.FG_RED_BOLD);
     }
