+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/common/Example9_1.java
@@ -1,28 +1,29 @@
-package com.rakovets.course.javabasics.example.concurrency;
+package com.rakovets.course.javabasics.example.concurrency.common;
 
 import com.rakovets.course.javabasics.util.AnsiColorCode;
 import com.rakovets.course.javabasics.util.StandardOutputUtil;
 
-public class Example10_1 {
+public class Example9_1 {
     public static void main(String[] args) {
         StandardOutputUtil.printlnWithTimeAndThread("started", AnsiColorCode.FG_RED_BOLD);
 
         Runnable runnable = () -> {
-            for (int i = 0; i < 5; i++) {
+            while (true) {
                 StandardOutputUtil.printlnWithTimeAndThread("...", AnsiColorCode.FG_BLUE_BOLD);
-                sleep(1000);
+                sleep(1);
             }
         };
         Thread thread = new Thread(runnable);
-        thread.setDaemon(true);
         thread.start();
+
+        sleep(5);
 
         StandardOutputUtil.printlnWithTimeAndThread("finished", AnsiColorCode.FG_RED_BOLD);
     }
 
-    private static void sleep(long millis) {
+    private static void sleep(long seconds) {
         try {
-            Thread.sleep(millis);
+            Thread.sleep(seconds * 1000);
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
