+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/common/Example9_2.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.javabasics.example.concurrency;
+package com.rakovets.course.javabasics.example.concurrency.common;
 
 import com.rakovets.course.javabasics.util.AnsiColorCode;
 import com.rakovets.course.javabasics.util.StandardOutputUtil;
