+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/common/model/Stoppable.java
@@ -1,0 +1,6 @@
+package com.rakovets.course.javabasics.example.concurrency.common.model;
+
+
+public interface Stoppable {
+    void stop();
+}
