--- src/main/java/com/rakovets/course/javabasics/example/concurrency/model/Stoppable.java
@@ -1,6 +1,0 @@
-package com.rakovets.course.javabasics.example.concurrency.model;
-
-
-public interface Stoppable {
-    void stop();
-}
