+++ src/main/java/com/rakovets/course/javabasics/example/concurrency/threadsynchronization/model/Store.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.javabasics.example.concurrent.model;
+package com.rakovets.course.javabasics.example.concurrency.threadsynchronization.model;
 
 import com.rakovets.course.javabasics.util.AnsiColorCode;
 import com.rakovets.course.javabasics.util.StandardOutputUtil;
